rename to cdq-analysis, use quay.io image
