+++ config/manager/kustomization.yaml
@@ -29,5 +29,5 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: controller
-  newName: yangcao77/hasoperator
-  newTag: "95"
+  newName: quay.io/redhat-appstudio/application-service
+  newTag: next
